+++ Modules/BulkPrint/Database/Migrations/2022_01_28_075807_create_fees_invoice_settings_table.php
@@ -47,10 +47,10 @@ class CreateFeesInvoiceSettingsTable extends Migration
 
             $table->text('copy_write_msg',255)->nullable();
 
-            $table->integer('created_by')->nullable()->default(1)->unsigned();
+            $table->unsignedBigInteger('created_by')->nullable()->default(1)->unsigned();
             $table->foreign('created_by')->references('id')->on('users')->onDelete('cascade');
 
-            $table->integer('updated_by')->nullable()->default(1)->unsigned();
+            $table->unsignedBigInteger('updated_by')->nullable()->default(1)->unsigned();
             $table->foreign('updated_by')->references('id')->on('users')->onDelete('cascade');
 
             $table->integer('school_id')->nullable()->default(1)->unsigned();
