+++ Modules/RolePermission/Database/Migrations/2020_01_11_141636_create_infix_module_infos_table.php
@@ -27,10 +27,10 @@ class CreateInfixModuleInfosTable extends Migration
             $table->string('lang_name')->nullable();
             $table->string('icon_class')->nullable();
             $table->tinyInteger('active_status')->default(1);
-            $table->integer('created_by')->nullable()->default(1)->unsigned();
+            $table->unsignedBigInteger('created_by')->nullable()->default(1)->unsigned();
             $table->foreign('created_by')->references('id')->on('users')->onDelete('cascade');
 
-            $table->integer('updated_by')->nullable()->default(1)->unsigned();
+            $table->unsignedBigInteger('updated_by')->nullable()->default(1)->unsigned();
             $table->foreign('updated_by')->references('id')->on('users')->onDelete('cascade');
 
             $table->integer('school_id')->nullable()->unsigned();
