+++ database/migrations/2020_01_28_103859_create_sm_news_pages_table.php
@@ -27,10 +27,10 @@ class CreateSmNewsPagesTable extends Migration
         $table->string('button_url')->nullable();
         $table->tinyInteger('active_status')->default(1);
 
-        $table->integer('created_by')->nullable()->default(1)->unsigned();
+        $table->unsignedBigInteger('created_by')->nullable()->default(1)->unsigned();
         $table->foreign('created_by')->references('id')->on('users')->onDelete('cascade');
 
-        $table->integer('updated_by')->nullable()->default(1)->unsigned();
+        $table->unsignedBigInteger('updated_by')->nullable()->default(1)->unsigned();
         $table->foreign('updated_by')->references('id')->on('users')->onDelete('cascade');
 
         $table->integer('school_id')->nullable()->default(1)->unsigned();
